exercises/serializers.py
# serializers.py
from rest_framework import serializers

from .models import Question

class QuestionSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Question
        fields = [f.name for f in Question._meta.get_fields()]