requirements.txt
Django==2.1.1
djangorestframework==3.11.0
numpy==1.17.4
pandas==0.25.3
psycopg2==2.8.4
